passed every restauration block into a functino to get a cleaner code
